Move x86/arm64 syscall hooks to corresponding modules
