+++ crates/unicorn/src/arch/arm64.rs
@@ -1,10 +1,13 @@
+use alloc::{boxed::Box, rc::Rc};
 use unicorn_engine_sys::{
-    Mode, RegisterARM64, RegisterARM64CP, uc_error, uc_reg_read, uc_reg_write,
+    Arm64Insn, HookType, Mode, RegisterARM64, RegisterARM64CP, uc_error, uc_hook_add, uc_reg_read,
+    uc_reg_write,
 };
 
 use crate::{
-    Unicorn,
+    UcHookId, Unicorn,
     arch::{Register, UcArch},
+    hook,
 };
 
 pub enum Arm64 {}
@@ -14,7 +17,7 @@ impl_reg_pc_counter!(RegisterARM64);
 
 // todo: find out if coprocessor functions can fail if the arch is correct
 // if they are infallible, Result is not needed
-impl<D> Unicorn<'_, D, Arm64> {
+impl<'a, D> Unicorn<'a, D, Arm64> {
     /// Read ARM64 Coprocessor register
     pub fn reg_read_arm64_coproc(&self, reg: &mut RegisterARM64CP) -> Result<(), uc_error> {
         unsafe {
@@ -37,5 +40,44 @@ impl<D> Unicorn<'_, D, Arm64> {
             )
         }
         .and(Ok(()))
+    }
+
+    /// Add hook for ARM MRS/MSR/SYS/SYSL instructions.
+    ///
+    /// If the callback returns true, the read/write to system registers would be skipped (even
+    /// though that may cause exceptions!). Note one callback per instruction is allowed.
+    pub fn add_insn_sys_hook<F>(
+        &mut self,
+        insn_type: Arm64Insn,
+        begin: u64,
+        end: u64,
+        callback: F,
+    ) -> Result<UcHookId, uc_error>
+    where
+        F: FnMut(&mut Unicorn<D, Arm64>, RegisterARM64, &RegisterARM64CP) -> bool + 'a,
+    {
+        let mut hook_id = 0;
+        let mut user_data = Box::new(hook::UcHook {
+            callback,
+            uc: Rc::downgrade(&self.inner),
+        });
+
+        unsafe {
+            uc_hook_add(
+                self.get_handle(),
+                (&raw mut hook_id).cast(),
+                HookType::INSN.0 as i32,
+                hook::insn_sys_hook_proxy_arm64::<D, F, Arm64> as _,
+                core::ptr::from_mut(user_data.as_mut()).cast(),
+                begin,
+                end,
+                insn_type,
+            )
+        }
+        .and_then(|| {
+            let hook_id = UcHookId(hook_id);
+            self.inner_mut().hooks.push((hook_id, user_data));
+            Ok(hook_id)
+        })
     }
 }
